+++ R/utils_assert.R
@@ -463,9 +463,11 @@ tar_assert_target <- function(x, msg = NULL) {
 #' @rdname tar_assert
 tar_assert_target_list <- function(x) {
   msg <- paste(
-    "The target script file (default: _targets.R)",
-    "must end with a list of tar_target() objects (recommended)",
-    "or a tar_pipeline() object (deprecated). Each element of the target list",
+    "Expected a list of target objects but did not find one.",
+    "Are you missing a target list at the end of your target script file?",
+    "The target script file (e.g. _targets.R)",
+    "must end with a list of tar_target() objects.",
+    "Each element of the target list",
     "must be a target object or nested list of target objects."
   )
   tar_assert_list(x, msg = msg)