+++ NEWS.md
@@ -9,6 +9,7 @@
 * Improve `tidyselect` docs in `tar_make()` (#640, @dewoller).
 * Use namespaced call to `tar_dir()` in `tar_test()` (#642, @billdenney).
 
+
 # targets 0.8.0
 
 ## Bug fixes