+++ NEWS.md
@@ -8,7 +8,7 @@
 
 * Improve `tidyselect` docs in `tar_make()` (#640, @dewoller).
 * Use namespaced call to `tar_dir()` in `tar_test()` (#642, @billdenney).
-
+* Improve `tar_assert_target_list()` error message (@kkami1115, #654).
 
 # targets 0.8.0